+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+* Replaced `-g none` with `-g native`
+
 ## [0.1.3] - 2022-10-27
 
 ### Updated